requirements.txt
MetaTrader5==5.0.4500
google-genai==0.3.0
pandas==2.2.0
pandas-ta==0.3.14b0
python-dotenv==1.0.1
pytz==2024.1